Create script to create nuget package.
While 'dotnet pack' is a cool and easy way to create nuget packages, it doesn't work so well for build tools, as dependencies are listed as nuget package dependencies, and it appears to be impossible to have the assemblies embedded in the nupkg. So, a msbuild file is created that will use the .NET SDK to create the nuget package, but only uses the auto-generated nuspec file, which is modifies and then re-packs to create the final nupkg.