Build solution that contains adfc projects
For some reason MSBuild behaves differently, with regards to automatic properties, when building a single project directly, or building a project as part of a solution. This commit adds a props file to nuget package, that adds the properties that do not get added automatically by the common msbuild targets file when building a solution.